+++ app/Http/Controllers/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\User;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 
@@ -15,6 +16,7 @@ class UserController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('isAdmin'), 404);
         $users = User::all();
         return view('user.index', compact('users'));
     }
@@ -26,6 +28,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('isAdmin'), 404);
         $user = new User();
         return view('user.create', compact('user'));
     }
@@ -38,6 +41,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(Gate::denies('isAdmin'), 404);
         $request->validate([
             'name' => 'required',
             'email' => 'required|email|unique:users,email',
@@ -56,6 +60,7 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        abort_if(Gate::denies('isAdmin'), 404);
         return view('user.show', compact('user'));
     }
 
@@ -67,6 +72,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        abort_if(Gate::denies('isAdmin'), 404);
         return view('user.edit', compact('user'));
     }
 
@@ -79,6 +85,7 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
+        abort_if(Gate::denies('isAdmin'), 404);
         $request->validate([
             'name' => 'required',
             'email' => 'required|email|unique:users,email,' . $user->id,
@@ -96,6 +103,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        abort_if(Gate::denies('isAdmin'), 404);
         $user->delete();
         return redirect()->route('user.index')->with('success', 'User has been deleted!');
     }
